Lighting_System: Added structure, classes, init, and lighting room bounds calculation
Basically by calculating the the range and intensity of the light
sources we can apply these values programmatically to the LightTiles
within the room. We know how far away the LightTiles are from the
sources because of the the Vector2 key being used to identify the
LightTile in the TileManager dictionaries.

The light ranges will also have calculations for light ‘fall off’ at
certain distances so that light tiles are not just  on or off but they
will gradually decline in alpha value the further away from the source.

At the moment LightTiles do not do anything at all. Will add that
functionality next..